Update logic to be more date focused.
This still uses the "seed" logic to generate a puzzle for a date,
but changes it to only have one puzzle per date, and otherwise
you generally have a fixed "seed" in the URL that updates whenever
you do a new game. If you have a "seed" that takes priority,
and if you specify a specific date, that is next, otherwise the
date is used to generate the word.